Always escalate exceptions in data generation
